flutter/pyproject.toml
[project]
name = "flutter"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "pillow>=12.1.0",
    "pyyaml>=6.0.3",
]
